9-08-4-Mark

+++ src/main/java/com/tencent/wxcloudrun/model/Mark.java
@@ -6,8 +6,8 @@ import java.math.BigDecimal;
 
 @Data
 public class Mark {
-    private Long markId; // 标记id
-    private Long id; // 房源id
+    private Integer markId; // 标记id
+    private Integer id; // 房源id
     private BigDecimal longitude; // 经度(地图选择上传)
     private BigDecimal latitude;  // 纬度(地图选择上传)
 }
